+++ src/vue/App.vue
@@ -125,7 +125,9 @@ export default {
 
     this.isAppInitialized = true
     Bus.success('change-password-form.password-changed-msg')
-    setTimeout(() => Bus.warning('change-password-form.password-changed-msg'), 2000)
+    setTimeout(() => Bus.warning('change-password-form.password-changed-msg'), 5000)
+    setTimeout(() => Bus.info('change-password-form.password-changed-msg'), 10000)
+    setTimeout(() => Bus.error('change-password-form.password-changed-msg'), 15000)
   },
 
   methods: {